+++ perforce/api.py
@@ -10,7 +10,7 @@ This module implements the Perforce API
 :license: MIT, see LICENSE for more details
 """
 
-from .models import Connection
+from .models import Connection, Client, Changelist
 from . import errors